Fix compiler warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@27128 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/apps/gdal_grid.cpp
@@ -263,7 +263,7 @@ static void ProcessCommonGeometry(OGRGeometry* poGeom, OGRGeometry *poClipSrc,
         {
             OGRLineString *poLS = (OGRLineString*)poGeom;
             OGRPoint point;
-            for (size_t pointIndex = 0; pointIndex < poLS->getNumPoints(); pointIndex++)
+            for (int pointIndex = 0; pointIndex < poLS->getNumPoints(); pointIndex++)
             {
                 poLS->getPoint(pointIndex, &point);
                 ProcessCommonGeometry((OGRGeometry*)&point, poClipSrc,
